musicbot/requirements.txt
# Core Dependencies
pyrogram>=2.0.0
tgcrypto>=1.2.0
pytgcalls>=3.0.0

# Audio Processing
yt-dlp>=2023.07.06
ffmpeg-python>=0.2.0

# Async Utilities
aiohttp>=3.8.0
aiofiles>=23.0.0
asyncio-mqtt>=0.11.0

# Database (Optional but Recommended)
motor>=3.0.0
pymongo>=4.0.0

# Cache Layer (Optional)
redis>=4.0.0

# Utilities
python-dotenv>=0.19.0
requests>=2.28.0
pillow>=9.0.0

# Logging and Monitoring
loguru>=0.6.0